Move Filter to lib
